fix: register_consumer
currently the implementation is non-functional with the current API implementation.
Two fixes are done in this PR.
First: the `Emails` field supplied with the API when empty should be removed from the payload.
Currently, if not given (as within the README), it will result in a empty object '{}'. The fix removes said object.

The second fix lies in the fetch of the consumer_id that wasn't done by symbol and crashes at runtime.