fix: credit order parsing
Currently all responses are loaded as symbols (see JSON parsing in http.rb).
Thus all map accesses through strings are currently failing.

This fixes the parsing methods to support symbols and correctly handle API responses.
This also adds a return value to the POST order call, returning the order_id, necessary to use
during the fetch of the credit order data.

note: the public method get_decision is currently not working though it is not in the path of
a simple integration to the API, it should be fixed next.